Disentangle some client/server dependencies.
