+++ server/src/time-types.ts
@@ -1,0 +1,13 @@
+export interface CurrentDelta {
+  delta: number;
+  dut1: number[] | null;
+  pendingLeap: number;
+  pendingLeapDate: string;
+}
+
+export interface TimeInfo {
+  time: number;
+  leapSecond: number;
+  leapExcess: number;
+  text: string;
+}
