+++ src/app.service.ts
@@ -1,7 +1,7 @@
 
 import { CurrentTemperatureHumidity } from './current-temp-manager';
 import { Settings } from './settings';
-import { TimeInfo } from '../server/src/time-poller';
+import { TimeInfo } from '../server/src/time-types';
 
 export interface AppService {
   forecastHasBeenUpdated(): void;
