+++ src/clock.ts
@@ -19,12 +19,12 @@
 
 // Started by using https://codepen.io/dudleystorey/pen/HLBki, but this has grown and changed *quite* a bit from there.
 
-import { DateAndTime, getDayOfWeek, getLastDateInMonthGregorian, KsDateTime, KsTimeZone } from 'ks-date-time-zone';
-import { isIE, isRaspbian, padLeft } from 'ks-util';
 import { AppService } from './app.service';
-import { CurrentDelta } from '../server/src/tai-utc';
 import * as $ from 'jquery';
+import { DateAndTime, getDayOfWeek, getLastDateInMonthGregorian, KsDateTime, KsTimeZone } from 'ks-date-time-zone';
 import { interpolate } from 'ks-math';
+import { isIE, isRaspbian, padLeft } from 'ks-util';
+import { CurrentDelta } from '../server/src/time-types';
 
 const SVG_NAMESPACE = 'http://www.w3.org/2000/svg';
